requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cobble==0.1.3
colorama==0.4.4
colorclass==2.2.2
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.11.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
docopt==0.6.2
future==0.18.2
html5lib==1.1
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==8.3.0
platformdirs==2.4.1
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.18.0
pytz==2021.3
reportlab==3.6.5
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
xhtml2pdf==0.2.5
